feat: add mobile and web stories for workout tracking panel
- Introduced `Mobile.stories.tsx` to showcase the workout tracking panel in a mobile viewport, illustrating various states such as NoBlock, ReadyToStart, ActiveFran, and AmrapRunning.
- Added `Web.stories.tsx` to demonstrate the web-based workout tracking panel with real ScriptRuntime, covering states like NoBlock, ReadyToStart, ActiveFran, ActiveAmrap, ActiveRounds, and ActiveEmom.
- Both stories utilize a real JIT-compiled workout script and provide a user interface for tracking workouts with appropriate controls and visual state panels.
