+++ stories/catalog/pages/Calendar.stories.tsx
@@ -10,7 +10,7 @@ import { Edit, BarChart2, Eye } from 'lucide-react';
 import React from 'react';
 
 const meta: Meta<typeof CalendarPageShell> = {
-  title: 'panels/Calendar',
+  title: 'catalog/pages/Calendar',
   component: CalendarPageShell,
   parameters: {
     layout: 'fullscreen',
