+++ stories/catalog/pages/JournalPageShell.stories.tsx
@@ -8,10 +8,10 @@
 
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
-import { StorybookWorkbench } from '../_shared/StorybookWorkbench';
+import { StorybookWorkbench } from '../../_shared/StorybookWorkbench';
 
 const meta: Meta = {
-  title: 'panels/JournalPageShell',
+  title: 'catalog/pages/JournalPageShell',
   parameters: {
     layout: 'fullscreen',
     docs: {
