+++ stories/catalog/templates/NoteEditor/Mobile.stories.tsx
@@ -8,10 +8,10 @@
 
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
-import { StorybookWorkbench } from '../../_shared/StorybookWorkbench';
+import { StorybookWorkbench } from '../../../_shared/StorybookWorkbench';
 
 const meta: Meta = {
-  title: 'Panels/NoteEditor/Mobile',
+  title: 'catalog/templates/NoteEditor/Mobile',
   decorators: [
     (Story) => (
       <div style={{ width: '390px', margin: '0 auto' }}>
