+++ stories/catalog/templates/NoteEditor/Web.stories.tsx
@@ -8,10 +8,10 @@
 
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
-import { StorybookWorkbench } from '../../_shared/StorybookWorkbench';
+import { StorybookWorkbench } from '../../../_shared/StorybookWorkbench';
 
 const meta: Meta = {
-  title: 'Panels/NoteEditor/Web',
+  title: 'catalog/templates/NoteEditor/Web',
   parameters: {
     layout: 'fullscreen',
     docs: {
